requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
pgvector==0.2.4
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# AWS
boto3==1.34.20
aioboto3==12.3.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
python-magic==0.4.27

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP & API
httpx==0.26.0
aiohttp==3.9.1
aiofiles==23.2.1

# Utilities
structlog==24.1.0
numpy==1.26.3

# Redis (optional)
redis==5.0.1

# Monitoring
prometheus-client==0.19.0